app/services/openai_service.py
import os
import openai
from dotenv import load_dotenv

load_dotenv()
openai.api_key = os.getenv("OPENAI_API_KEY")

async def generate_story(word: str) -> str:
    prompt = f"Write a short fantasy story that teaches the meaning of the word '{word}' through a magical tale."

    response = openai.ChatCompletion.create(
        model="gpt-4",
        messages=[{"role": "user", "content": prompt}],
        temperature=0.7,
        max_tokens=500
    )
    return response.choices[0]["message"]["content"]
