requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==1.109.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
